[Refactor] design-tokens.ts를 admin-design-tokens.ts로 rename
- design-tokens.ts → admin-design-tokens.ts 파일명 변경
- 관련 import 경로 7개 파일 수정
- index.css 주석 업데이트
- CLAUDE.md 프로젝트 가이드 문서 업데이트
