+++ src/components/layout/ta/TenantAdminLayout.tsx
@@ -1,7 +1,7 @@
 import { useState, type ReactNode } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { TenantAdminSidebar } from './TenantAdminSidebar';
-import { designTokens } from '@/styles/design-tokens';
+import { designTokens } from '@/styles/admin-design-tokens';
 import { tenantAdminMenuData } from '@/config/sidebar-menus';
 
 interface TenantAdminLayoutProps {
