+++ src/components/layout/to/TenantOperatorLayout.tsx
@@ -1,7 +1,7 @@
 import { useState, type ReactNode } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { TenantOperatorSidebar } from './TenantOperatorSidebar';
-import { designTokens } from '@/styles/design-tokens';
+import { designTokens } from '@/styles/admin-design-tokens';
 import { tenantOperatorMenuData } from '@/config/sidebar-menus';
 
 interface TenantOperatorLayoutProps {
